+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+### Changed
+- Title background
 ### Added
+- Custom ONT/OND css for reports
+- Custom ONT/OND color classes
 - Added ability to turn off about section
 
 ## [v0.6.3]